fix: react does not recognize the duotoneColor prop on a DOM element. Closes #65
