+++ package.json
@@ -133,5 +133,6 @@
         }
       ]
     ]
-  }
+  },
+  "packageManager": "yarn@1.22.21+sha1.1959a18351b811cdeedbd484a8f86c3cc3bbaf72"
 }
